Extract functions for interacting with SC API
