+++ pivotal-import/api.py
@@ -1,0 +1,36 @@
+import os
+
+import requests
+
+# API Helpers
+sc_token = os.getenv("SHORTCUT_API_TOKEN")
+api_url_base = "https://api.app.shortcut.com/api/v3"
+headers = {
+    "Shortcut-Token": sc_token,
+    "Accept": "application/json; charset=utf-8",
+    "Content-Type": "application/json",
+}
+
+
+def sc_get(path, params={}):
+    url = api_url_base + path
+    logger.debug("GET url=%s params=%s headers=%s" % (url, params, headers))
+    resp = requests.get(url, headers=headers, params=params)
+    resp.raise_for_status()
+    return resp.json()
+
+
+def sc_post(path, data={}):
+    url = api_url_base + path
+    logger.debug("POST url=%s params=%s headers=%s" % (url, data, headers))
+    resp = requests.post(url, headers=headers, json=data)
+    resp.raise_for_status()
+    return resp.json()
+
+
+def sc_put(path, data={}):
+    url = api_url_base + path
+    logger.debug("PUT url=%s params=%s headers=%s" % (url, data, headers))
+    resp = requests.put(url, headers=headers, json=data)
+    resp.raise_for_status()
+    return resp.json()
